Introduce a PromotionHandler namespace
There we can put classes resposible for triggering promotion logic in
specific spots of the code base. e.g. Cart when item is added to cart,
Coupon when customer submits a coupon code
