+++ core/lib/spree/core.rb
@@ -42,7 +42,6 @@ require 'spree/core/engine'
 
 require 'spree/i18n'
 require 'spree/money'
-require 'spree/promo/coupon_applicator'
 
 require 'spree/core/delegate_belongs_to'
 require 'spree/core/permalinks'
